+++ toolbus-ng-argus-viewer/src/argusviewer/view/architectureview/performance/tree/ProcessorRenderer.java
@@ -18,7 +18,7 @@ public class ProcessorRenderer implements TableCellRenderer {
 	 */
 	public Component getTableCellRendererComponent(JTable table, Object value,
 			boolean isSelected, boolean hasFocus, int row, int column) {
-		Integer processor = (Integer) value;
+		int processor = ((Integer) value).intValue();
 
 		JProgressBar progressBar = new JProgressBar();
 		progressBar.setValue(processor);
